moved test classes into task package
