+++ src/test/java/dolla/task/DebtTest.java
@@ -1,6 +1,5 @@
-package dolla;
+package dolla.task;
 
-import dolla.task.Debt;
 import org.junit.jupiter.api.Test;
 import java.time.LocalDate;
